chore: add produce & do chores
