+++ .node-version
@@ -1,1 +1,1 @@
-18.16.0
+20.11.0
