+++ package.json
@@ -3,18 +3,9 @@
   "version": "0.0.19",
   "private": true,
   "type": "module",
-  "packageManager": "pnpm@8.6.7",
-  "homepage": "https://github.com/sheinsight/reactive#readme",
+  "packageManager": "pnpm@8.15.3",
   "author": {
     "name": "ityuany"
-  },
-  "license": "MIT",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/sheinsight/reactive.git"
-  },
-  "bugs": {
-    "url": "https://github.com/sheinsight/reactive/issues"
   },
   "scripts": {
     "build": "pnpm --filter '@shined/*' build ",
@@ -24,20 +15,18 @@
     "release": "tsx ./scripts/release.ts"
   },
   "devDependencies": {
-    "@types/node": "^20.4.5",
-    "@types/semver": "^7.5.0",
+    "@types/node": "^20.11.19",
+    "@types/semver": "^7.5.7",
     "@vitest/coverage-v8": "^0.33.0",
-    "enquirer": "^2.3.6",
-    "execa": "^7.1.1",
+    "enquirer": "^2.4.1",
+    "execa": "^7.2.0",
     "find-packages": "^10.0.4",
-    "happy-dom": "^10.3.2",
-    "read-pkg": "^8.0.0",
+    "happy-dom": "^10.11.2",
+    "read-pkg": "^8.1.0",
     "read-yaml-file": "^2.1.0",
-    "semver": "^7.5.4",
-    "tsx": "^3.12.7",
+    "semver": "^7.6.0",
+    "tsx": "^3.14.0",
     "vitest": "^0.33.0",
-    "write-pkg": "^6.0.0"
-  },
-  "readme": "ERROR: No README data found!",
-  "_id": "reactive@0.0.18"
+    "write-pkg": "^6.0.1"
+  }
 }
