+++ packages/reactive/package.json
@@ -1,10 +1,26 @@
 {
   "name": "@shined/reactive",
   "version": "0.0.19",
-  "description": "A reactive react state management library based on Proxy",
+  "type": "module",
   "main": "dist/index.js",
   "module": "dist/index.js",
-  "type": "module",
+  "description": "A reactive react state management library based on Proxy",
+  "homepage": "https://github.com/sheinsight/reactive#readme",
+  "author": {
+    "name": "ityuany"
+  },
+  "files": [
+    "./dist",
+    "./src"
+  ],
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/sheinsight/reactive.git",
+    "directory": "packages/reactive"
+  },
+  "bugs": {
+    "url": "https://github.com/sheinsight/reactive/issues"
+  },
   "scripts": {
     "dev": "tsup --watch",
     "build": "tsup",
@@ -12,17 +28,10 @@
     "test:w": "vitest",
     "test:cov": "vitest run --coverage"
   },
-  "files": [
-    "./dist",
-    "./hack-remove-readonly.d.ts"
-  ],
   "exports": {
     ".": {
       "import": "./dist/index.js",
       "default": "./dist/index.js"
-    },
-    "./hack-remove-readonly": {
-      "types": "./hack-remove-readonly.d.ts"
     },
     "./package.json": "./package.json"
   },
@@ -31,7 +40,6 @@
     "registry": "https://registry.npmjs.org/"
   },
   "sideEffects": false,
-  "packageManager": "pnpm@8.6.7",
   "keywords": [
     "react",
     "reactive",
@@ -41,32 +49,27 @@
     "use-sync-external-store",
     "proxy-compare"
   ],
-  "author": {
-    "name": "ityuany"
-  },
   "license": "MIT",
   "dependencies": {
-    "@redux-devtools/extension": "^3.2.5",
+    "@redux-devtools/extension": "^3.3.0",
     "@types/use-sync-external-store": "^0.0.3",
-    "proxy-compare": "^2.5.1",
+    "proxy-compare": "^2.6.0",
     "use-sync-external-store": "^1.2.0"
   },
   "devDependencies": {
-    "@swc/core": "^1.3.78",
+    "@swc/core": "^1.4.2",
     "@testing-library/react": "^12.1.5",
     "@testing-library/react-hooks": "^8.0.1",
-    "@types/node": "^20.4.5",
+    "@types/node": "^20.11.19",
     "@types/react": "17.0.62",
     "@vitest/coverage-v8": "^0.33.0",
     "react": "^16",
     "react-dom": "^16.8.6",
     "react-test-renderer": "^16.8.6",
     "tsup": "^7.1.0",
-    "typescript": "^5.0.2"
+    "typescript": "^5.3.3"
   },
   "peerDependencies": {
     "react": "^16.9.0 || ^17.0.0 || ^18.0.0"
-  },
-  "readme": "ERROR: No README data found!",
-  "_id": "@shined/reactive@0.0.18"
+  }
 }
