--- packages/reactive/src/original.ts
@@ -1,6 +1,0 @@
-import { ORIGINAL } from "./utils.js";
-
-/** @deprecated */
-export function original<T>(object: T) {
-  return object[ORIGINAL]?.() as T;
-}
