+++ packages/reactive/src/snapshot.ts
@@ -1,0 +1,5 @@
+import { SNAPSHOT } from "./internal-utils.js";
+
+export function getSnapshot<T extends object>(proxyState: T): T {
+  return proxyState[SNAPSHOT];
+}
