+++ packages/reactive/src/subscribe.spec.ts
@@ -1,6 +1,7 @@
-import { subscribe } from "./subscribe.js";
-import { LISTENERS } from "./utils.js";
 import { describe, it, expect, vitest } from "vitest";
+
+import { subscribe } from "./subscribe.js";
+import { LISTENERS } from "./internal-utils.js";
 
 describe("subscribe", () => {
   it("should add a listener to proxyObject", () => {
