+++ packages/reactive/tsup.config.ts
@@ -2,8 +2,11 @@ import { defineConfig } from "tsup";
 
 export default defineConfig({
   target: "es5",
-  entry: ["src/index.ts"],
-  splitting: false,
+  /**
+   * @issue https://github.com/egoist/tsup/issues/1000
+   */
+  entry: ["src/index.ts", "src/utils.ts"], // export `./utils` entry in `package.json`
+  splitting: true,
   sourcemap: true,
   clean: true,
   dts: true,
